requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.0
neo4j==5.14.1
pandas==2.1.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
